refactor(tidy): recover high-value runtime surfaces checkpoint
