+++ tests/Integration/test_panel_auth_shell_coherence.py
@@ -8,7 +8,8 @@ from fastapi import HTTPException
 
 from backend.web.models.panel import PublishAgentRequest, UpdateAgentRequest
 from backend.web.routers import panel as panel_router
-from backend.web.services import agent_user_service, library_service
+from backend.threads import agent_user_service
+from backend.library import service as library_service
 from storage.contracts import UserRow, UserType
