+++ tests/Integration/test_resource_overview_contract_split.py
@@ -8,7 +8,7 @@ from fastapi.testclient import TestClient
 
 import backend.sandboxes.resources.projection as resource_projection_service
 import backend.sandboxes.resources.provider_boundary as resource_provider_boundary_service
-from backend import resource_common
+from backend.sandboxes.resources import common as resource_common
 from backend.monitor.api.http import router as monitor_router
 from backend.monitor.infrastructure.read_models import resource_read_service as monitor_resource_read_service
 from backend.monitor.infrastructure.web import gateway as monitor_gateway
