+++ tests/Integration/test_sandbox_router_user_shell.py
@@ -5,7 +5,7 @@ from types import SimpleNamespace
 import pytest
 
 from backend.web.routers import sandbox as sandbox_router
-from backend.web.services import sandbox_service
+from backend.sandboxes import service as sandbox_service
 
 
 @pytest.mark.asyncio
