+++ tests/Integration/test_schema_redesign_baseline_contract.py
@@ -5,12 +5,10 @@ from types import SimpleNamespace
 
 import backend.sandboxes.resources.projection as resource_projection_service
 import backend.sandboxes.resources.provider_boundary as resource_provider_boundary_service
-from backend import resource_common
-from backend import resource_provider_contracts as resource_service
+from backend.sandboxes.resources import common as resource_common
+from backend.sandboxes import service as sandbox_service
+from backend.sandboxes.resources import provider_contracts as resource_service
 from backend.monitor.infrastructure.read_models import resource_read_service as monitor_resource_read_service
-from backend.web.services import (
-    sandbox_service,
-)
 from backend.web.utils.serializers import avatar_url
