--- tests/Unit/backend/web/services/test_resource_common.py
@@ -1,303 +1,0 @@
-import inspect
-
-import pytest
-
-from backend import resource_common, resource_projection, user_resource_projection
-from backend import resource_common as neutral_resource_common
-from backend import resource_io as neutral_resource_io
-from backend import resource_io as resource_io
-from backend import resource_provider_boundary as neutral_resource_provider_boundary
-from backend import resource_provider_contracts as neutral_resource_provider_contracts
-from backend import resource_provider_contracts as resource_contracts
-from backend import sandbox_provider_factory as neutral_sandbox_provider_factory
-from backend import user_sandbox_reads as neutral_user_sandbox_reads
-from backend.web.services import (
-    account_resource_service,
-    sandbox_service,
-)
-
-
-class _FakeThreadRepo:
-    def __init__(self, rows):
-        self._rows = rows
-
-    def get_by_id(self, thread_id: str):
-        return self._rows.get(thread_id)
-
-    def list_by_ids(self, thread_ids: list[str]):
-        return [{"id": thread_id, **row} for thread_id, row in self._rows.items() if thread_id in set(thread_ids)]
-
-    def close(self):
-        pass
-
-
-class _FakeAgent:
-    def __init__(self, user_id: str, display_name: str, avatar: str | None = None):
-        self.id = user_id
-        self.display_name = display_name
-        self.avatar = avatar
-
-
-class _FakeUserRepo:
-    def __init__(self, users):
-        self._users = users
-
-    def list_all(self):
-        return list(self._users)
-
-    def close(self):
-        pass
-
-
-@pytest.mark.parametrize(
-    ("thread_repo", "expected_thread_ids"),
-    [
-        (_FakeThreadRepo({"thread-1": {"agent_user_id": "agent-1"}}), ["thread-1", "thread-2"]),
-    ],
-)
-def test_thread_owners_resolve_runtime_storage_metadata(thread_repo, expected_thread_ids):
-    owners = resource_common.thread_owners(
-        expected_thread_ids,
-        thread_repo=thread_repo,
-        user_repo=_FakeUserRepo([_FakeAgent("agent-1", "Toad", avatar="x")]),
-    )
-
-    assert owners == {
-        "thread-1": {"agent_user_id": "agent-1", "agent_name": "Toad", "avatar_url": "/api/users/agent-1/avatar"},
-        "thread-2": {"agent_user_id": None, "agent_name": "未绑定Agent", "avatar_url": None},
-    }
-
-
-def test_thread_owners_prefers_batch_thread_lookup() -> None:
-    class _BatchOnlyThreadRepo:
-        def list_by_ids(self, thread_ids: list[str]):
-            assert thread_ids == ["thread-1", "thread-2"]
-            return [{"id": "thread-1", "agent_user_id": "agent-1"}]
-
-        def get_by_id(self, _thread_id: str):
-            raise AssertionError("unexpected per-thread lookup")
-
-        def close(self):
-            pass
-
-    owners = resource_common.thread_owners(
-        ["thread-1", "thread-2"],
-        thread_repo=_BatchOnlyThreadRepo(),
-        user_repo=_FakeUserRepo([_FakeAgent("agent-1", "Toad", avatar="x")]),
-    )
-
-    assert owners == {
-        "thread-1": {"agent_user_id": "agent-1", "agent_name": "Toad", "avatar_url": "/api/users/agent-1/avatar"},
-        "thread-2": {"agent_user_id": None, "agent_name": "未绑定Agent", "avatar_url": None},
-    }
-
-
-@pytest.mark.parametrize(
-    ("thread_repo", "user_repo", "message"),
-    [
-        (
-            type(
-                "_BrokenThreadRepo",
-                (),
-                {
-                    "list_by_ids": lambda self, _thread_ids: (_ for _ in ()).throw(RuntimeError("thread repo offline")),
-                    "close": lambda self: None,
-                },
-            )(),
-            _FakeUserRepo([_FakeAgent("agent-1", "Toad", avatar="x")]),
-            "thread repo offline",
-        ),
-        (
-            _FakeThreadRepo({"thread-1": {"agent_user_id": "agent-1"}}),
-            type(
-                "_BrokenUserRepo",
-                (),
-                {
-                    "list_all": lambda self: (_ for _ in ()).throw(RuntimeError("user repo offline")),
-                    "close": lambda self: None,
-                },
-            )(),
-            "user repo offline",
-        ),
-    ],
-    ids=["thread-repo", "user-repo"],
-)
-def test_thread_owners_fail_loudly_on_repo_errors(thread_repo, user_repo, message) -> None:
-    with pytest.raises(RuntimeError, match=message):
-        resource_common.thread_owners(
-            ["thread-1"],
-            thread_repo=thread_repo,
-            user_repo=user_repo,
-        )
-
-
-def test_metric_adds_error_only_when_present():
-    assert resource_common.metric(1, 2, "%", "api", "live") == {
-        "used": 1,
-        "limit": 2,
-        "unit": "%",
-        "source": "api",
-        "freshness": "live",
-    }
-    assert resource_common.metric(None, None, "GB", "unknown", "stale", "probe failed") == {
-        "used": None,
-        "limit": None,
-        "unit": "GB",
-        "source": "unknown",
-        "freshness": "stale",
-        "error": "probe failed",
-    }
-
-
-def test_resource_metrics_normalize_live_snapshot_values() -> None:
-    metrics = resource_common.to_resource_metrics(
-        {
-            "cpu_used": 25,
-            "memory_used_mb": 2048,
-            "memory_total_mb": 4096,
-            "disk_used_gb": 12,
-            "disk_total_gb": 50,
-            "network_rx_kbps": 100,
-            "network_tx_kbps": 50,
-        }
-    )
-
-    assert metrics == {
-        "cpu": 25.0,
-        "memory": 2.0,
-        "memoryLimit": 4.0,
-        "memoryNote": None,
-        "disk": 12.0,
-        "diskLimit": 50.0,
-        "diskNote": None,
-        "networkIn": 100.0,
-        "networkOut": 50.0,
-        "probeError": None,
-    }
-
-
-def test_shared_resource_consumers_use_neutral_resource_common_owner() -> None:
-    projection_source = inspect.getsource(resource_projection)
-    resource_io_source = inspect.getsource(resource_io)
-
-    assert "backend.web.services.resource_common" not in projection_source
-    assert "backend.web.services.resource_common" not in resource_io_source
-    assert "backend.sandboxes.resources.common" in projection_source
-
-
-def test_resource_common_owner_keeps_expected_surface() -> None:
-    assert resource_common.metric is neutral_resource_common.metric
-    assert resource_common.thread_owners is neutral_resource_common.thread_owners
-
-
-def test_shared_resource_consumers_use_neutral_resource_provider_boundary_owner() -> None:
-    projection_source = inspect.getsource(resource_projection)
-    user_projection_source = inspect.getsource(user_resource_projection)
-
-    assert "backend.web.services import resource_provider_boundary_service" not in projection_source
-    assert "backend.web.services import resource_provider_boundary_service" not in user_projection_source
-    assert "backend.sandboxes.resources.provider_boundary" in projection_source
-    assert "backend.sandboxes.resources.provider_boundary" in user_projection_source
-
-
-def test_resource_provider_boundary_uses_neutral_sandbox_inventory_owner() -> None:
-    boundary_source = inspect.getsource(neutral_resource_provider_boundary)
-
-    assert "sandbox_service.available_sandbox_types" not in boundary_source
-    assert "backend.sandboxes.inventory" in boundary_source
-
-
-def test_resource_provider_boundary_moves_user_sandbox_reads_out_of_sandbox_service() -> None:
-    boundary_source = inspect.getsource(neutral_resource_provider_boundary)
-    sandbox_service_source = inspect.getsource(sandbox_service)
-
-    assert "sandbox_service.list_user_sandboxes" not in boundary_source
-    assert "backend.sandboxes.user_reads" in boundary_source
-    assert "user_sandbox_reads.list_user_sandboxes(" in sandbox_service_source
-    assert "user_sandbox_reads._list_user_runtime_rows(" in sandbox_service_source
-
-
-def test_user_sandbox_reads_uses_neutral_avatar_and_virtual_thread_helpers() -> None:
-    source = inspect.getsource(neutral_user_sandbox_reads)
-
-    assert "backend.web.utils.serializers" not in source
-    assert "backend.web.utils.helpers" not in source
-    assert "backend.web.services.thread_visibility" not in source
-    assert "backend.identity.avatar.urls" in source
-    assert "backend.threads.virtual_threads" in source
-    assert "backend.threads.projection" in source
-    assert "def count_user_visible_sandboxes_by_provider(" in source
-
-
-def test_sandbox_service_uses_neutral_helper_owners_for_read_wrappers() -> None:
-    source = inspect.getsource(sandbox_service)
-
-    assert "backend.web.services.thread_visibility" not in source
-    assert "backend.web.utils.serializers" not in source
-    assert "backend.web.utils.helpers" not in source
-    assert "backend.threads.projection" in source
-    assert "backend.identity.avatar.urls" in source
-    assert "backend.threads.virtual_threads" in source
-
-
-def test_account_resource_service_uses_neutral_sandbox_count_owner() -> None:
-    source = inspect.getsource(account_resource_service)
-
-    assert "backend.web.services import sandbox_service" not in source
-    assert "user_sandbox_reads" in source
-
-
-def test_resource_modules_use_neutral_sandbox_path_owner() -> None:
-    common_source = inspect.getsource(neutral_resource_common)
-    projection_source = inspect.getsource(resource_projection)
-
-    assert "backend.web.core.config" not in common_source
-    assert "backend.web.core.config" not in projection_source
-    assert "backend.sandboxes.paths" in common_source
-    assert "backend.sandboxes.paths" in projection_source
-
-
-def test_resource_modules_use_neutral_sandbox_provider_factory_owner() -> None:
-    common_source = inspect.getsource(neutral_resource_common)
-    resource_io_source = inspect.getsource(resource_io)
-    sandbox_provider_factory_source = inspect.getsource(neutral_sandbox_provider_factory)
-
-    assert "backend.sandboxes.service" not in common_source
-    assert "backend.sandboxes.service" not in resource_io_source
-    assert "backend.sandboxes.service" not in sandbox_provider_factory_source
-    assert "backend.sandboxes.provider_factory" in common_source
-    assert "backend.sandboxes.provider_factory" in resource_io_source
-    assert "backend.sandboxes.inventory" in sandbox_provider_factory_source
-
-
-def test_sandbox_provider_factory_returns_none_without_filesystem_probe(monkeypatch) -> None:
-    monkeypatch.setattr(
-        "backend.sandboxes.inventory.init_providers_and_managers",
-        lambda: ({}, {}),
-    )
-
-    assert not hasattr(neutral_sandbox_provider_factory, "SANDBOXES_DIR")
-    assert neutral_sandbox_provider_factory.build_provider_from_config_name("missing-provider") is None
-
-
-def test_web_sandbox_service_keeps_provider_factory_compat_surface() -> None:
-    assert sandbox_service.build_provider_from_config_name is neutral_sandbox_provider_factory.build_provider_from_config_name
-
-
-def test_resource_io_owner_moves_out_of_resource_service() -> None:
-    io_source = inspect.getsource(neutral_resource_io)
-
-    assert "backend.web.services import resource_service" not in io_source
-    assert resource_io.refresh_resource_snapshots is neutral_resource_io.refresh_resource_snapshots
-
-
-def test_resource_provider_boundary_moves_display_contract_out_of_resource_service() -> None:
-    boundary_source = inspect.getsource(neutral_resource_provider_boundary)
-
-    assert "backend.web.services import resource_service" not in boundary_source
-    assert "resource_provider_contracts" in boundary_source
-
-
-def test_resource_provider_contract_owner_keeps_expected_surface() -> None:
-    assert resource_contracts.get_provider_display_contract is neutral_resource_provider_contracts.get_provider_display_contract
-    assert resource_contracts.build_resource_row_payload is neutral_resource_provider_contracts.build_resource_row_payload
