+++ tests/Unit/sandbox/test_sandbox_service_cleanup.py
@@ -1,6 +1,6 @@
 from types import SimpleNamespace
 
-from backend.web.services import sandbox_service
+from backend.sandboxes import service as sandbox_service
 
 LOWER_RUNTIME_KEY = "lease_" + "id"
