+++ tests/Unit/sandbox/test_sandbox_user_runtime_rows.py
@@ -2,7 +2,7 @@ from types import SimpleNamespace
 
 import pytest
 
-from backend.web.services import sandbox_service
+from backend.sandboxes import service as sandbox_service
 
 LOWER_RUNTIME_KEY = "lease_" + "id"
